# Conflicts:
#	src/loadjpk.py

+++ src/loadjpk.py
@@ -15,17 +15,11 @@ from jpkfile import JPKFile,JPKMap
 import zipfile
 from zipfile import ZipFile
 from scipy.signal import savgol_filter
-import multiprocessing
+from nanoscope import files
+from nanoscope.constants import FORCE, METRIC, VOLTS, PLT_kwargs
 def rotate(data_x,data_y,index,k):
     theta = np.arctan(k)*-1
     return (data_x - data_x[index])*np.sin(theta) + (data_y - data_y[index])*np.cos(theta) +data_y[index]
-def writedatay(ljp,index,todir):
-    data = ljp[index]
-    datamsg =  ljp.data_lst[index]
-    filename = "{}-{}.datay".format(os.path.splitext(os.path.basename(datamsg[0]))[0],datamsg[1])
-    fname = os.path.join(todir,filename)
-    with open(fname,'wb') as f:
-            pickle.dump(data,f)
 class forcecurve:
     def __init__(self):
         self.data = {'tasktype':'',
@@ -58,6 +52,7 @@ class forcecurve:
             if tip_correc:
                 data[k]['measuredHeight'] = data[k]['measuredHeight'] - data[k]['vDeflection']/self.data['springConstant']
             if 'k' in self.data['offset'].keys():
+                print('ok')
                 data[k]['vDeflection'] = rotate(data[k]['measuredHeight'].reshape(-1),data[k]['vDeflection'].reshape(-1),-1,self.data['offset']['k']).reshape(-1,1)
         return data
     def savedata2txt(self,savedir='data.txt'):
@@ -126,14 +121,18 @@ class loadjpkfile(forcecurve):
                         for i in range(maxindex):
                             self.datalst.append((fname,i))
                         break
-                
+            elif sum([True for i in ['.spm'] if fname.endswith(i)]):
+                with files.ForceVolumeFile(fname) as f:
+                    fv_pixels = f.force_curves_channel.number_of_force_curves
+                    for i in range(fv_pixels):
+                        self.datalst.append((fname,i))
     def get_filenamelst(self,Travel=True):
         if os.path.isfile(self.filedir):
             self.filelst.append(self.filedir)
         elif os.path.isdir(self.filedir):
             for a,b,c in os.walk(self.filedir, topdown=True, onerror=None, followlinks=False):
                 for filename in c:
-                    if sum([True for i in ['.txt','.jpk-force','.jpk-force-map','.datay'] if os.path.join(a, filename).endswith(i)]):
+                    if sum([True for i in ['.txt','.jpk-force','.jpk-force-map','.datay','spm'] if os.path.join(a, filename).endswith(i)]):
                         self.filelst.append(os.path.join(a, filename))
                 if not Travel:
                     break
@@ -146,6 +145,8 @@ class loadjpkfile(forcecurve):
             self.extract_map_data(filename,index)
         elif filename.endswith('.datay'):
             self.extract_datay_data(filename,index)
+        elif filename.endswith('.spm'):
+            self.extract_spm_data(filename,index)
     def extract_txt_data(self,filename,index):
         data = np.loadtxt(filename,comments='#')
         with open(filename,'r') as f:
@@ -176,26 +177,57 @@ class loadjpkfile(forcecurve):
         self.data['springConstant'] = springConstant
         for i,segment in jpk.segments.items():
             self.data['rawdata'][segment.get_info('type')] = segment.get_array(['measuredHeight','vDeflection'])[0]
+    def extract_spm_data(self,fname,index):
+        with files.ForceVolumeFile(fname) as f:
+            fc_channel = f.force_curves_channel
+            h_sens_chan = f[2]
+            fz_plot, ax_prop = fc_channel.create_force_z_plot(index, FORCE)
+            h_sens_data = h_sens_chan.get_force_curve_data(index, METRIC)
+            if 'nN' in ax_prop['ylabel']:
+                factor = 1e-9
+            elif 'pN' in ax_prop['ylabel']:
+                factor = 1e-12
+            data_x=h_sens_data.retrace*-1e-9
+            data_y=fz_plot.retrace.y*factor
+            data = np.dstack((data_x,data_y))[0]
+            self.data['rawdata']['retract'] = np.array([[tuple(i)] for i in data],dtype=[('measuredHeight', '<f8'), ('vDeflection', '<f8')])
+            self.data['springConstant'] = f.spring_constant
+    def extract_all_map2datay(self,todir):
+        dic = self.data_structure
+        for filename in self.filelst:
+            print(filename)
+            if filename.endswith('.jpk-force-map'):
+                jpks = JPKMap(filename)
+                for i in range(len(jpks.flat_indices)):
+
+                    jpk = jpks.get_single_pixel(i)
+                    try:
+                        springConstant = float(jpk.shared_parameters['lcd-info']['2']['conversion-set']['conversion']['force']['scaling']['multiplier'])
+                    except:
+                        springConstant = float(jpk.shared_parameters['lcd-info']['1']['conversion-set']['conversion']['force']['scaling']['multiplier'])
+                    dic['springConstant'] = springConstant
+                    dic['datamsg']=(filename,i)
+                    for n,segment in jpk.segments.items():
+                        dic['rawdata'][segment.get_info('type')] = segment.get_array(['measuredHeight','vDeflection'])[0]
+                    fname = "{}-{}.datay".format(os.path.join(todir,os.path.splitext(os.path.basename(filename))[0]),i)
+                    with open(fname,'wb') as f:
+                        pickle.dump(dic,f)
+            elif filename.endswith('.jpk-force-map'):
+                jpk = JPKFile(filename)
+                try:
+                    springConstant = float(jpk.shared_parameters['lcd-info']['2']['conversion-set']['conversion']['force']['scaling']['multiplier'])
+                except:
+                    continue
+                dic['springConstant'] = springConstant
+                fname = "{}-{}.datay".format(os.path.splitext(os.path.basename(filename))[0],0)
+                with open(fname,'wb') as f:
+                    pickle.dump(dic,f)
     def extract_datay_data(self,filename,index):
         with open(filename,'rb') as f:
             pkl = pickle.load(f)
         springConstant = pkl['springConstant']
         self.data['springConstant'] = springConstant
         self.data['rawdata'] = pkl['rawdata']
-    def wriredatay(self,todir,index):
-        data = self[index]
-        datamsg =  self.data_lst[index]
-        filename = "{}-{}.datay".format(os.path.splitext(os.path.basename(datamsg[0]))[0],datamsg[1])
-        fname = os.path.join(todir,filename)
-        with open(fname,'wb') as f:
-            pickle.dump(data,f)
-    def conver_jpk_datay(self,todir):
-        if os.path.isdir(todir):
-            pool = multiprocessing.Pool(processes = 3)
-            for i in range(len(self)):
-                pool.apply_async(self.wriredatay, (todir, i))
-            pool.close()
-            pool.join()
 class zipfileopera:
     def __init__(self,fname='test.DataYee-force'):
         self.fname = fname
